rockshop-app/embed-svc/requirements.txt
torch
torchvision
pillow
scikit-learn
tabulate
flask
